move countries data fetch and storage to container
