+++ src/scenes/country-scene/country-scene.js
@@ -2,17 +2,31 @@ import React, { useEffect, useState } from 'react';
 import { loadModules } from 'esri-loader';
 import { connect } from 'react-redux';
 import Component from './country-scene-component';
-import actions from 'redux_modules/country-extent';
+import countryExtentActions from 'redux_modules/country-extent';
+import countryDataActions from 'redux_modules/country-data';
 
-import { COUNTRIES_GENERALIZED_BORDERS_FEATURE_LAYER } from 'constants/layers-slugs';
+import {
+  COUNTRIES_GENERALIZED_BORDERS_FEATURE_LAYER,
+  COUNTRIES_DATA_FEATURE_LAYER
+ } from 'constants/layers-slugs';
 import { LAYERS_URLS } from 'constants/layers-urls';
+import mapStateToProps from './country-scene-selectors';
 
-
+const actions = {...countryExtentActions, ...countryDataActions }
 const CountrySceneContainer = (props) => {
-
-  const { countryISO, setCountryExtentLoading, setCountryExtentReady, setCountryExtentError } = props;
+  const {
+    countriesData,
+    countryISO,
+    setCountryExtentLoading,
+    setCountryExtentReady,
+    setCountryExtentError,
+    setCountryDataLoading,
+    setCountryDataReady,
+    setCountryDataError
+  } = props;
 
   const [countryLayer, setCountryLayer] = useState(null);
+  const [countriesDataLayer, setCountriesDataLayer] = useState(null);
 
   const createExtent = () => {
     const query = countryLayer.createQuery();
@@ -37,8 +51,29 @@ const CountrySceneContainer = (props) => {
         url: LAYERS_URLS[COUNTRIES_GENERALIZED_BORDERS_FEATURE_LAYER]
       });
       setCountryLayer(_countryLayer)
+      const _countriesDataLayer = new FeatureLayer({
+        url: LAYERS_URLS[COUNTRIES_DATA_FEATURE_LAYER]
+      });
+      _countriesDataLayer.outFields = ['*'];
+      setCountriesDataLayer(_countriesDataLayer)
     });
   }, []);
+
+
+  useEffect(() => {
+    if (countriesDataLayer && !countriesData) {
+      setCountryDataLoading();
+      const query = countriesDataLayer.createQuery();
+      countriesDataLayer.queryFeatures(query)
+      .then((results) => {
+        const { features } = results;
+        setCountryDataReady(features);
+      })
+      .catch((error) => {
+        setCountryDataError(error);
+      });
+    }
+  }, [countriesDataLayer, countriesData])
 
   useEffect(() => {
     if (countryLayer && countryISO) {
@@ -52,4 +87,4 @@ const CountrySceneContainer = (props) => {
   )
 }
 
-export default connect(null, actions)(CountrySceneContainer);
+export default connect(mapStateToProps, actions)(CountrySceneContainer);
